pyproject.toml
[build-system]
requires = [
    "setuptools<65",
    "lsst-versions >= 1.3.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "rubintv_analysis_service"
description = "A set of libraries and scripts for running the backend services for the rubin tv visual framework."
version = "0.0.1"
license = {text = "GPLv3+ License"}
readme = "README.md"
authors = [
    {name="Rubin Observatory Data Management", email="dm-admin@lists.lsst.org"},
]
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Astronomy",
]
keywords = ["lsst", "visualization"]
dependencies = [
    "numpy>=1.25.2",
    "scipy",
    "matplotlib",
    "pydantic",
    "pyyaml",
    "sqlalchemy",
    "astropy",
    "websocket-client",
    "lsst-daf-butler",
    # temporary dependency for testing
    "tornado",
]
#dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/lsst-ts/rubintv_analysis_service"

[project.optional-dependencies]
test = [
    "pytest >= 3.2",
    "flake8 >= 3.7.5",
    "pytest-flake8 >= 1.0.4",
]

[tool.setuptools.packages.find]
where = ["python"]

[tool.setuptools]
zip-safe = false
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = { attr = "lsst_versions.get_lsst_version" }

[tool.black]
line-length = 110
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 110

[tool.lsst_versions]
write_to = "python/lsst/rubintv/analysis/service/version.py"

# The matplotlib test may not release font files.
open_files_ignore = ["*.ttf"]

[tool.pydocstyle]
convention = "numpy"
# Our coding style does not require docstrings for magic methods (D105)
# Our docstyle documents __init__ at the class level (D107)
# We allow methods to inherit docstrings and this is not compatible with D102.
# Docstring at the very first line is not required
# D200, D205 and D400 all complain if the first sentence of the docstring does
# not fit on one line.
add-ignore = ["D107", "D105", "D102", "D100", "D200", "D205", "D400"]
